requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==24.3.0
boto3==1.35.82
botocore==1.35.82
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
fastapi==0.104.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.12
langchain-aws==0.2.9
langchain-community
langchain-core
langchain-text-splitters==0.3.3
langsmith==0.2.3
mangum==0.17.0
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.12
packaging==23.2
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.4
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.27.0
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.0.7
uvicorn==0.24.0
wheel==0.44.0
yarl==1.18.3
